+++ sanity/schemas/index.ts
@@ -1,5 +1,7 @@
 // documents
 import site from './documents/site'
+import header from './documents/header'
+import footer from './documents/footer'
 import page from './documents/page'
 import redirect from './documents/redirect'
 import blogPost from './documents/blog.post'
@@ -22,6 +24,8 @@ import heroPostcard from './modules/hero.postcard'
 export const schemaTypes = [
 	// documents
 	site,
+	header,
+	footer,
 	page,
 	redirect,
 	blogPost,